+++ database/demo-en/myems_system_db.sql
@@ -80,7 +80,7 @@ VALUES
 (1, 'inlet chilled water temperature', 'point', NULL, 2, NULL, NULL),
 (1, 'chilled_water instantaneous flow rate', 'point', NULL, 3, NULL, NULL),
 (1, 'instantaneous power', 'point', NULL, 4, NULL, NULL),
-(1, 'COP', 'fraction', NULL, NULL, '5ca47bc5-22c2-47fc-b906-33222191ea40', '719acf65-b932-4176-94e5-e9fe420abc68');
+(1, 'COP', 'fraction', NULL, NULL, '5ca47bc5-22c2-47fc-b906-33222191ea40', '6db58cd6-33d3-58ed-a095-22333202fb51');
 
 COMMIT;
 
@@ -360,7 +360,7 @@ VALUES
 (1, 'inlet chilled water temperature', 'point', NULL, 2, NULL, NULL),
 (1, 'chilled_water instantaneous flow rate', 'point', NULL, 3, NULL, NULL),
 (1, 'instantaneous power', 'point', NULL, 4, NULL, NULL),
-(1, 'COP', 'fraction', NULL, NULL, '5ca47bc5-22c2-47fc-b906-33222191ea40', '719acf65-b932-4176-94e5-e9fe420abc68');
+(1, 'COP', 'fraction', NULL, NULL, '5ca47bc5-22c2-47fc-b906-33222191ea40', '6db58cd6-33d3-58ed-a095-22333202fb51');
 
 COMMIT;