+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - updated distribution system export, import and clone functions in myems-api
 - updated meter export, import and clone functions in myems-api
 - updated equipment table in myems-admin
+- updated equipment export, import and clone functions in myems-api
+- 
 ### Fixed
 - updated run action of rule in myems-api
 - updated run action of advanced report in myems-api